Remove response multimethod, use jedis-inspired bulk response reader
Moving to a single method should be better for performance - multimethod
flexibility really isn't necessary for this.

The existing bulk method parsing code used .read to read into a byte array.
It turns out that in practice, this doesn't always fill the array, leaving bytes
on the wire. This is moderately disastrous. Take a page from Jedis' parsing
code and use .read in a loop to make sure we get all of the content.